requirements.txt
azure-functions
playwright
